+++ src/client.rs
@@ -11,19 +11,14 @@ use tokio_postgres::{
 /// Not every **pgstac** function is provided, and some names are changed to
 /// match Rust conventions.
 #[derive(Debug)]
-pub struct Client<C>(C)
+pub struct Client<'a, C>(&'a C)
 where
     C: GenericClient;
 
-impl<C: GenericClient> Client<C> {
+impl<'a, C: GenericClient> Client<'a, C> {
     /// Creates a new client.
-    pub fn new(client: C) -> Client<C> {
+    pub fn new(client: &C) -> Client<C> {
         Client(client)
-    }
-
-    /// Returns this client's inner client.
-    pub fn into_inner(self) -> C {
-        self.0
     }
 
     /// Returns the **pgstac** version.
@@ -110,8 +105,8 @@ impl<C: GenericClient> Client<C> {
         self.value("search", &[&search]).await
     }
 
-    async fn query_one<'a>(
-        &'a self,
+    async fn query_one(
+        &self,
         function: &str,
         params: &[&(dyn ToSql + Sync)],
     ) -> std::result::Result<Row, tokio_postgres::Error> {
@@ -208,17 +203,17 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn version(client: &Client<Transaction<'_>>) {
+    async fn version(client: &Client<'_, Transaction<'_>>) {
         let _ = client.version().await.unwrap();
     }
 
     #[pgstac_test]
-    async fn setting(client: &Client<Transaction<'_>>) {
+    async fn setting(client: &Client<'_, Transaction<'_>>) {
         assert_eq!(client.setting("context").await.unwrap(), "off");
     }
 
     #[pgstac_test]
-    async fn collections(client: &Client<Transaction<'_>>) {
+    async fn collections(client: &Client<'_, Transaction<'_>>) {
         assert!(client.collections().await.unwrap().is_empty());
         client
             .add_collection(Collection::new("an-id", "a description"))
@@ -228,7 +223,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn add_collection_duplicate(client: &Client<Transaction<'_>>) {
+    async fn add_collection_duplicate(client: &Client<'_, Transaction<'_>>) {
         assert!(client.collections().await.unwrap().is_empty());
         let collection = Collection::new("an-id", "a description");
         client.add_collection(collection.clone()).await.unwrap();
@@ -236,7 +231,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn upsert_collection(client: &Client<Transaction<'_>>) {
+    async fn upsert_collection(client: &Client<'_, Transaction<'_>>) {
         assert!(client.collections().await.unwrap().is_empty());
         let mut collection = Collection::new("an-id", "a description");
         client.upsert_collection(collection.clone()).await.unwrap();
@@ -255,7 +250,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn update_collection(client: &Client<Transaction<'_>>) {
+    async fn update_collection(client: &Client<'_, Transaction<'_>>) {
         let mut collection = Collection::new("an-id", "a description");
         client.add_collection(collection.clone()).await.unwrap();
         assert!(client
@@ -281,18 +276,18 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn update_collection_does_not_exit(client: &Client<Transaction<'_>>) {
+    async fn update_collection_does_not_exit(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("an-id", "a description");
         assert!(client.update_collection(collection).await.is_err());
     }
 
     #[pgstac_test]
-    async fn collection_not_found(client: &Client<Transaction<'_>>) {
+    async fn collection_not_found(client: &Client<'_, Transaction<'_>>) {
         assert!(client.collection("not-an-id").await.unwrap().is_none());
     }
 
     #[pgstac_test]
-    async fn delete_collection(client: &Client<Transaction<'_>>) {
+    async fn delete_collection(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("an-id", "a description");
         client.add_collection(collection.clone()).await.unwrap();
         assert!(client.collection("an-id").await.unwrap().is_some());
@@ -301,12 +296,12 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn delete_collection_does_not_exist(client: &Client<Transaction<'_>>) {
+    async fn delete_collection_does_not_exist(client: &Client<'_, Transaction<'_>>) {
         assert!(client.delete_collection("not-an-id").await.is_err());
     }
 
     #[pgstac_test]
-    async fn item(client: &Client<Transaction<'_>>) {
+    async fn item(client: &Client<'_, Transaction<'_>>) {
         assert!(client
             .item("an-id", "collection-id")
             .await
@@ -329,13 +324,13 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn item_without_collection(client: &Client<Transaction<'_>>) {
+    async fn item_without_collection(client: &Client<'_, Transaction<'_>>) {
         let item = Item::new("an-id");
         assert!(client.add_item(item.clone()).await.is_err());
     }
 
     #[pgstac_test]
-    async fn update_item(client: &Client<Transaction<'_>>) {
+    async fn update_item(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("collection-id", "a description");
         client.add_collection(collection).await.unwrap();
         let mut item = Item::new("an-id");
@@ -359,7 +354,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn upsert_item(client: &Client<Transaction<'_>>) {
+    async fn upsert_item(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("collection-id", "a description");
         client.add_collection(collection).await.unwrap();
         let mut item = Item::new("an-id");
@@ -370,7 +365,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn add_items(client: &Client<Transaction<'_>>) {
+    async fn add_items(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("collection-id", "a description");
         client.add_collection(collection).await.unwrap();
         let mut item = Item::new("an-id");
@@ -392,7 +387,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn upsert_items(client: &Client<Transaction<'_>>) {
+    async fn upsert_items(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("collection-id", "a description");
         client.add_collection(collection).await.unwrap();
         let mut item = Item::new("an-id");
@@ -406,7 +401,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn search_everything(client: &Client<Transaction<'_>>) {
+    async fn search_everything(client: &Client<'_, Transaction<'_>>) {
         assert!(client
             .search(Search::default())
             .await
@@ -426,7 +421,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn search_ids(client: &Client<Transaction<'_>>) {
+    async fn search_ids(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("collection-id", "a description");
         client.add_collection(collection).await.unwrap();
         let mut item = Item::new("an-id");
@@ -446,7 +441,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn search_collections(client: &Client<Transaction<'_>>) {
+    async fn search_collections(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("collection-id", "a description");
         client.add_collection(collection).await.unwrap();
         let mut item = Item::new("an-id");
@@ -466,7 +461,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn search_limit(client: &Client<Transaction<'_>>) {
+    async fn search_limit(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("collection-id", "a description");
         client.add_collection(collection).await.unwrap();
         let mut item = Item::new("an-id");
@@ -485,7 +480,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn search_bbox(client: &Client<Transaction<'_>>) {
+    async fn search_bbox(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("collection-id", "a description");
         client.add_collection(collection).await.unwrap();
         let mut item = Item::new("an-id");
@@ -505,7 +500,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn search_datetime(client: &Client<Transaction<'_>>) {
+    async fn search_datetime(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("collection-id", "a description");
         client.add_collection(collection).await.unwrap();
         let mut item = Item::new("an-id");
@@ -526,7 +521,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn search_intersects(client: &Client<Transaction<'_>>) {
+    async fn search_intersects(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("collection-id", "a description");
         client.add_collection(collection).await.unwrap();
         let mut item = Item::new("an-id");
@@ -558,7 +553,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn pagination(client: &Client<Transaction<'_>>) {
+    async fn pagination(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("collection-id", "a description");
         client.add_collection(collection).await.unwrap();
         let mut item = Item::new("an-id");
@@ -584,7 +579,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn fields(client: &Client<Transaction<'_>>) {
+    async fn fields(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("collection-id", "a description");
         client.add_collection(collection).await.unwrap();
         let mut item = Item::new("an-id");
@@ -611,7 +606,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn sortby(client: &Client<Transaction<'_>>) {
+    async fn sortby(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("collection-id", "a description");
         client.add_collection(collection).await.unwrap();
         let mut item = Item::new("a");
@@ -644,7 +639,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn filter(client: &Client<Transaction<'_>>) {
+    async fn filter(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("collection-id", "a description");
         client.add_collection(collection).await.unwrap();
         let mut item = Item::new("a");
@@ -671,7 +666,7 @@ mod tests {
     }
 
     #[pgstac_test]
-    async fn query(client: &Client<Transaction<'_>>) {
+    async fn query(client: &Client<'_, Transaction<'_>>) {
         let collection = Collection::new("collection-id", "a description");
         client.add_collection(collection).await.unwrap();
         let mut item = Item::new("a");